backend-fastapi/vision/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
psycopg[binary]>=3.1.0
psycopg-pool>=3.2.0
numpy>=1.26.0
torch>=2.2.0
torchvision>=0.17.0
Pillow>=10.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pandas>=2.0.0
pyarrow>=14.0.0
